chore: add errors

+++ server/campus_hub/utils/db.py
@@ -4,6 +4,7 @@ from pymongo import MongoClient
 from pymongo.collection import Collection
 from pymongo.errors import ConnectionFailure, PyMongoError
 from werkzeug.exceptions import InternalServerError
+from campus_hub.utils.errors import error
 
 
 class DBConnector:
@@ -42,7 +43,9 @@ class DBConnector:
             self.client.admin.command("ping")
             return True
         except ConnectionFailure as e:
-            self.logger.error(f"Error pinging MongoDB: {e}")
+            self.logger.error(
+                error(500, "ServerError", "Error connecting to Mongo client", e)
+            )
             return False
 
     def close_connection(self) -> None:
@@ -53,7 +56,9 @@ class DBConnector:
             self.client.close()
             self.logger.info("MongoDB connection closed.")
         except PyMongoError as e:
-            self.logger.error(f"Error closing MongoDB connection: {e}")
+            self.logger.error(
+                error(500, "ServerError", "Error closing mongo connection.", e)
+            )
 
     def insert_data(self, collection_name: str, data: dict) -> None:
         """
@@ -75,7 +80,11 @@ class DBConnector:
             collection.insert_one(data)
             self.logger.info("Data inserted successfully.")
         except PyMongoError as e:
-            self.logger.error(f"Error inserting data into MongoDB: {e}")
+            self.logger.error(
+                error(
+                    500, "ServerError", "Error inserting data to Mongo collection.", e
+                )
+            )
             raise InternalServerError("Failed to insert data into MongoDB")
 
     def collection_exists(self, collection: Collection) -> bool:
@@ -100,7 +109,14 @@ class DBConnector:
         try:
             self.db.create_collection(collection_name)
         except PyMongoError as e:
-            self.logger.error(f"Error creating collection '{collection_name}': {e}")
+            self.logger.error(
+                error(
+                    500,
+                    "ServerError",
+                    f"Error creating {collection_name} collection.",
+                    e,
+                )
+            )
             raise InternalServerError(
                 f"Failed to create collection '{collection_name}'"
             )
@@ -121,7 +137,9 @@ class DBConnector:
             result = list(collection.find(query))
             return result
         except PyMongoError as e:
-            self.logger.error(f"Error querying data from MongoDB: {e}")
+            self.logger.error(
+                error(500, "ServerError", "Error querying data from MongoDB.", e)
+            )
             raise InternalServerError("Failed to query data from MongoDB")
 
     def update_data(self, collection_name: str, query: dict, update_data: dict) -> None:
@@ -135,10 +153,16 @@ class DBConnector:
         """
         try:
             collection = self.db[collection_name]
+
+            # Check if the document(s) to update exist
+            if collection.count_documents(query) == 0:
+                raise LookupError("No matching records found for the update")
+
+            # Update the data
             collection.update_many(query, {"$set": update_data})
             self.logger.info("Data updated successfully.")
         except PyMongoError as e:
-            self.logger.error(f"Error updating data in MongoDB: {e}")
+            self.logger.error(error(500, "ServerError", "Error updating data.", e))
             raise InternalServerError("Failed to update data in MongoDB")
 
     def delete_data(self, collection_name: str, query: dict) -> None:
@@ -151,10 +175,15 @@ class DBConnector:
         """
         try:
             collection = self.db[collection_name]
+
+            # Check if the document(s) to delete exist
+            if collection.count_documents(query) == 0:
+                raise LookupError("No matching records found to delete")
+
             collection.delete_many(query)
             self.logger.info("Data deleted successfully.")
         except PyMongoError as e:
-            self.logger.error(f"Error deleting data in MongoDB: {e}")
+            self.logger.error(error(500, "ServerError", "Error deleting data.", e))
             raise InternalServerError("Failed to delete data in MongoDB")
